Move all context providers to a withContexts HOC
