+++ src/app/containers/Article/index.jsx
@@ -4,6 +4,7 @@ import compose from '../../lib/utilities/compose';
 import articlePropTypes from '../../models/propTypes/article';
 import ArticleMain from '../ArticleMain';
 
+import withContexts from '../PageHandlers/withContexts';
 import withPageWrapper from '../PageHandlers/withPageWrapper';
 import withError from '../PageHandlers/withError';
 import withLoading from '../PageHandlers/withLoading';
@@ -22,6 +23,7 @@ ArticleContainer.defaultProps = {
 };
 
 const EnhancedArticleContainer = compose(
+  withContexts,
   withPageWrapper,
   withLoading,
   withError,
