+++ src/app/reducers/FeatureFlagReducer/index.jsx
@@ -1,8 +1,10 @@
 // FeatureFlag Actions
-const updateFeatureFlags = data => ({
-  type: 'UPDATE_FEATURE_FLAGS',
-  data,
-});
+// const updateFeatureFlags = data => ({
+//   type: 'UPDATE_FEATURE_FLAGS',
+//   data,
+// });
+
+const updateFeatureFlags = data => console.log('feature flags updating');
 
 const featureFlagReducer = (featureFlagState, action) => {
   const { data, type } = action;
